# Michel/improve performance