--- a/internal/base64_json/base64_json.go
+++ b/internal/base64_json/base64_json.go
@@ -4,6 +4,7 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"github.com/iancoleman/orderedmap"
+	"sort"
 )
 
 var (
@@ -22,8 +23,10 @@ var (
 	}
 )
 
-func keyOrder(p1, p2 *orderedmap.Pair) bool {
-	return keyIndexes[p1.Key()] < keyIndexes[p2.Key()]
+func keySort(keys []string) {
+	sort.Slice(keys, func(i, j int) bool {
+		return keyIndexes[keys[i]] < keyIndexes[keys[j]]
+	})
 }
 
 func Encode(v interface{}) (*string, error) {
@@ -45,7 +48,7 @@ func Encode(v interface{}) (*string, error) {
 			return nil, err
 		}
 	}
-	orderedMap.Sort(keyOrder)
+	orderedMap.SortKeys(keySort)
 	orderedJson, err := json.Marshal(orderedMap)
 	if err != nil {
 		return nil, err