--- a/internal/base64_json/base64_json.go
+++ b/internal/base64_json/base64_json.go
@@ -27,19 +27,23 @@ func keyOrder(p1, p2 *orderedmap.Pair) bool {
 }
 
 func Encode(v interface{}) (*string, error) {
-	// Note: []byte values are encoded as base64-encoded strings
-	//       (see: https://golang.org/pkg/encoding/json/#Marshal)
-	buf, err := json.Marshal(v)
-	if err != nil {
-		return nil, err
-	}
+	orderedMap, isOrderedMap := v.(*orderedmap.OrderedMap)
 
-	// Struct fields are marshalled in order of declaration, but we can't easily change the order
-	// OrderedMap fields are always marshalled in order, so we bounce through it
-	orderedMap := orderedmap.New()
-	err = json.Unmarshal(buf, orderedMap)
-	if err != nil {
-		return nil, err
+	if !isOrderedMap {
+		// Note: []byte values are encoded as base64-encoded strings
+		//       (see: https://golang.org/pkg/encoding/json/#Marshal)
+		buf, err := json.Marshal(v)
+		if err != nil {
+			return nil, err
+		}
+
+		// Struct fields are marshalled in order of declaration, but we can't easily change the order
+		// OrderedMap fields are always marshalled in order, so we bounce through it
+		orderedMap = orderedmap.New()
+		err = json.Unmarshal(buf, orderedMap)
+		if err != nil {
+			return nil, err
+		}
 	}
 	orderedMap.Sort(keyOrder)
 	orderedJson, err := json.Marshal(orderedMap)